+++ hil-test/tests/spi_full_duplex_dma_pcnt.rs
@@ -8,12 +8,8 @@
 use esp_hal::{
     dma::{Dma, DmaPriority, DmaRxBuf, DmaTxBuf},
     dma_buffers,
-    gpio::{Io, Pull},
-    pcnt::{
-        channel::{EdgeMode, PcntInputConfig, PcntSource},
-        unit::Unit,
-        Pcnt,
-    },
+    gpio::{interconnect::InputSignal, Io},
+    pcnt::{channel::EdgeMode, unit::Unit, Pcnt},
     peripherals::SPI2,
     prelude::*,
     spi::{
@@ -38,7 +34,7 @@ cfg_if::cfg_if! {
 
 struct Context {
     spi: SpiDma<'static, SPI2, DmaChannel0, FullDuplexMode, Blocking>,
-    pcnt_source: PcntSource,
+    pcnt_source: InputSignal,
     pcnt_unit: Unit<'static, 0>,
 }
 
@@ -79,7 +75,7 @@ mod tests {
 
         Context {
             spi,
-            pcnt_source: PcntSource::from(mosi_loopback_pcnt, PcntInputConfig { pull: Pull::Down }),
+            pcnt_source: mosi_loopback_pcnt,
             pcnt_unit: pcnt.unit0,
         }
     }